+++ src/components/datetime/DateTime.js
@@ -7,6 +7,7 @@ export class DateTimeComponent extends BaseComponent {
   constructor(component, options, data) {
     super(component, options, data);
     this.validators.push('date');
+    this.closedOn = 0;
   }
 
   elementInfo() {
@@ -18,26 +19,32 @@ export class DateTimeComponent extends BaseComponent {
     return info;
   }
 
-  build() {
-    super.build();
+  /**
+   * Get the default date for the calendar.
+   * @return {*}
+   */
+  get defaultDate() {
+    if (!this.component.defaultDate) {
+      return null;
+    }
 
-    // See if a default date is set.
-    if (this.component.defaultDate) {
-      var defaultDate = new Date(this.component.defaultDate);
-      if (!defaultDate || isNaN(defaultDate.getDate())) {
-        try {
-          let moment = momentModule;
-          defaultDate = new Date(eval(this.component.defaultDate));
-        }
-        catch (e) {
-          defaultDate = '';
-        }
+    let defaultDate = new Date(this.component.defaultDate);
+    if (!defaultDate || isNaN(defaultDate.getDate())) {
+      try {
+        let moment = momentModule;
+        defaultDate = new Date(eval(this.component.defaultDate));
       }
-
-      if (defaultDate && !isNaN(defaultDate.getDate())) {
-        this.setValue(defaultDate);
+      catch (e) {
+        defaultDate = null;
       }
     }
+
+    // Ensure this is a date.
+    if (defaultDate && isNaN(defaultDate.getDate())) {
+      defaultDate = null;
+    }
+
+    return defaultDate;
   }
 
   // This select component can handle multiple items on its own.
@@ -45,6 +52,11 @@ export class DateTimeComponent extends BaseComponent {
     return false;
   }
 
+  /**
+   * Convert the format from the angular-datepicker module.
+   * @param format
+   * @return {string|XML|*}
+   */
   convertFormat(format) {
     // Year conversion.
     format = format.replace(/y/g, 'Y');
@@ -84,10 +96,13 @@ export class DateTimeComponent extends BaseComponent {
       noCalendar: !_get(this.component, 'enableDate', true),
       altFormat: this.convertFormat(_get(this.component, 'format', '')),
       dateFormat: 'U',
-      defaultDate: _get(this.component, 'defaultDate', ''),
+      defaultDate: this.defaultDate,
       hourIncrement: _get(this.component, 'timePicker.hourStep', 1),
       minuteIncrement: _get(this.component, 'timePicker.minuteStep', 5),
-      onChange: () => this.onChange()
+      minDate: _get(this.component, 'datePicker.minDate'),
+      maxDate: _get(this.component, 'datePicker.maxDate'),
+      onChange: () => this.onChange(),
+      onClose: () => (this.closedOn = Date.now())
     };
   }
 
@@ -95,6 +110,12 @@ export class DateTimeComponent extends BaseComponent {
     super.disabled = disabled;
     _each(this.inputs, (input) => {
       if (input.calendar) {
+        if (disabled) {
+          input.calendar._input.setAttribute('disabled', 'disabled');
+        }
+        else {
+          input.calendar._input.removeAttribute('disabled');
+        }
         input.calendar.redraw();
       }
     });
@@ -102,9 +123,16 @@ export class DateTimeComponent extends BaseComponent {
 
   addSuffix(input, inputGroup) {
     let suffix = this.ce('span', {
-      class: 'input-group-addon'
+      class: 'input-group-addon',
+      style: 'cursor: pointer'
     });
     suffix.appendChild(this.getIcon(this.component.enableDate ? 'calendar' : 'time'));
+    this.addEventListener(suffix, 'click', () => {
+      // Make sure the calendar is not already open and that it did not just close (like from blur event).
+      if (!input.calendar.isOpen && ((Date.now() - this.closedOn) > 200)) {
+        input.calendar.open();
+      }
+    });
     inputGroup.appendChild(suffix);
     return suffix;
   }
@@ -117,8 +145,7 @@ export class DateTimeComponent extends BaseComponent {
   getDate(value) {
     let timestamp = parseInt(value, 10);
     if (!timestamp) {
-      // Just default to today.
-      return (new Date());
+      return null;
     }
     return (new Date(timestamp * 1000));
   }
@@ -135,13 +162,22 @@ export class DateTimeComponent extends BaseComponent {
   }
 
   getValueAt(index) {
-    return this.getDate(this.inputs[index].value).toISOString();
+    if (!this.inputs[index] || !this.inputs[index].calendar) {
+      return '';
+    }
+
+    let dates = this.inputs[index].calendar.selectedDates;
+    if (!dates || !dates.length) {
+      return '';
+    }
+
+    return dates[0].toISOString();
   }
 
   setValueAt(index, value) {
-    if (this.inputs[index].calendar) {
+    if (value && this.inputs[index].calendar) {
       let date = value ? new Date(value) : new Date();
-      this.inputs[index].calendar.setDate(date);
+      this.inputs[index].calendar.setDate(date, false);
     }
   }
 }